feat(generator): update dependencies and implement DID resolver caching with authentication
