+++ packages/generator/package.json
@@ -11,12 +11,12 @@
   },
   "dependencies": {
     "shared": "workspace:*",
-    "@atcute/atproto": "^3.0.3",
-    "@atcute/bluesky": "^3.0.4",
-    "@atcute/identity": "^1.0.2",
-    "@atcute/identity-resolver": "^1.1.3",
-    "@atcute/lexicons": "^1.0.4",
-    "@atcute/xrpc-server": "^0.1.1"
+    "@atcute/atproto": "^3.1.10",
+    "@atcute/bluesky": "^3.2.17",
+    "@atcute/identity": "^1.1.3",
+    "@atcute/identity-resolver": "^1.2.2",
+    "@atcute/lexicons": "^1.2.9",
+    "@atcute/xrpc-server": "^0.1.9"
   },
   "devDependencies": {
     "@cloudflare/vitest-pool-workers": "0.12.7",
