+++ packages/generator/src/auth/didResolverCache.ts
