--- vitest.workspace.ts
@@ -1,5 +1,0 @@
-export default [
-  './packages/generator/vitest.config.ts',
-  './packages/editor/vitest.config.ts',
-  './packages/shared/vitest.config.ts',
-];
